Hide inaccessible symbols in internal linkage.
