+++ src/AllocationBoundsInference.cpp
@@ -14,6 +14,8 @@ using std::set;
 using std::string;
 using std::vector;
 
+namespace {
+
 // Figure out the region touched of each buffer, and deposit them as
 // let statements outside of each realize node, or at the top level if
 // they're not internal allocations.
@@ -153,6 +155,8 @@ class StripDeclareBoxTouched : public IRMutator {
         }
     }
 };
+
+}  // namespace
 
 Stmt allocation_bounds_inference(Stmt s,
                                  const map<string, Function> &env,
