+++ src/CodeGen_Metal_Dev.cpp
@@ -2,6 +2,8 @@
 #include <sstream>
 #include <utility>
 
+#include "CodeGen_C.h"
+#include "CodeGen_GPU_Dev.h"
 #include "CodeGen_Internal.h"
 #include "CodeGen_Metal_Dev.h"
 #include "Debug.h"
@@ -16,6 +18,87 @@ using std::string;
 using std::vector;
 
 static ostringstream nil;
+
+namespace {
+
+class CodeGen_Metal_Dev : public CodeGen_GPU_Dev {
+public:
+    CodeGen_Metal_Dev(Target target);
+
+    /** Compile a GPU kernel into the module. This may be called many times
+     * with different kernels, which will all be accumulated into a single
+     * source module shared by a given Halide pipeline. */
+    void add_kernel(Stmt stmt,
+                    const std::string &name,
+                    const std::vector<DeviceArgument> &args) override;
+
+    /** (Re)initialize the GPU kernel module. This is separate from compile,
+     * since a GPU device module will often have many kernels compiled into it
+     * for a single pipeline. */
+    void init_module() override;
+
+    std::vector<char> compile_to_src() override;
+
+    std::string get_current_kernel_name() override;
+
+    void dump() override;
+
+    std::string print_gpu_name(const std::string &name) override;
+
+    std::string api_unique_name() override {
+        return "metal";
+    }
+
+protected:
+    class CodeGen_Metal_C : public CodeGen_C {
+    public:
+        CodeGen_Metal_C(std::ostream &s, Target t)
+            : CodeGen_C(s, t) {
+        }
+        void add_kernel(const Stmt &stmt,
+                        const std::string &name,
+                        const std::vector<DeviceArgument> &args);
+
+    protected:
+        using CodeGen_C::visit;
+        std::string print_type(Type type, AppendSpaceIfNeeded space_option = DoNotAppendSpace) override;
+        // Vectors in Metal come in two varieties, regular and packed.
+        // For storage allocations and pointers used in address arithmetic,
+        // packed types must be used. For temporaries, constructors, etc.
+        // regular types must be used.
+        // This concept also potentially applies to half types, which are
+        // often only supported for storage, not arithmetic,
+        // hence the method name.
+        std::string print_storage_type(Type type);
+        std::string print_type_maybe_storage(Type type, bool storage, AppendSpaceIfNeeded space);
+        std::string print_reinterpret(Type type, const Expr &e) override;
+        std::string print_extern_call(const Call *op) override;
+
+        std::string get_memory_space(const std::string &);
+
+        std::string shared_name;
+
+        void visit(const Min *) override;
+        void visit(const Max *) override;
+        void visit(const Div *) override;
+        void visit(const Mod *) override;
+        void visit(const For *) override;
+        void visit(const Ramp *op) override;
+        void visit(const Broadcast *op) override;
+        void visit(const Call *op) override;
+        void visit(const Load *op) override;
+        void visit(const Store *op) override;
+        void visit(const Select *op) override;
+        void visit(const Allocate *op) override;
+        void visit(const Free *op) override;
+        void visit(const Cast *op) override;
+        void visit(const Atomic *op) override;
+    };
+
+    std::ostringstream src_stream;
+    std::string cur_kernel_name;
+    CodeGen_Metal_C metal_c;
+};
 
 CodeGen_Metal_Dev::CodeGen_Metal_Dev(Target t)
     : metal_c(src_stream, t) {
@@ -724,6 +807,12 @@ void CodeGen_Metal_Dev::dump() {
 
 std::string CodeGen_Metal_Dev::print_gpu_name(const std::string &name) {
     return name;
+}
+
+}  // namespace
+
+CodeGen_GPU_Dev *new_CodeGen_Metal_Dev(const Target &target) {
+    return new CodeGen_Metal_Dev(target);
 }
 
 }  // namespace Internal
