+++ src/Memoization.cpp
@@ -311,8 +311,6 @@ public:
     }
 };
 
-}  // namespace
-
 // Inject caching structure around memoized realizations.
 class InjectMemoization : public IRMutator {
 public:
@@ -436,6 +434,8 @@ private:
     }
 };
 
+}  // namespace
+
 Stmt inject_memoization(const Stmt &s, const std::map<std::string, Function> &env,
                         const std::string &name,
                         const std::vector<Function> &outputs) {
@@ -449,6 +449,8 @@ Stmt inject_memoization(const Stmt &s, const std::map<std::string, Function> &en
 
     return injector.mutate(s);
 }
+
+namespace {
 
 class RewriteMemoizedAllocations : public IRMutator {
 public:
@@ -537,6 +539,8 @@ private:
         }
     }
 };
+
+}  // namespace
 
 Stmt rewrite_memoized_allocations(const Stmt &s, const std::map<std::string, Function> &env) {
