+++ src/ScheduleFunctions.h
@@ -12,9 +12,11 @@
 #include <vector>
 
 #include "Expr.h"
-#include "Target.h"
 
 namespace Halide {
+
+struct Target;
+
 namespace Internal {
 
 class Function;
