+++ src/StrictifyFloat.cpp
@@ -7,6 +7,8 @@
 namespace Halide {
 namespace Internal {
 
+namespace {
+
 class StrictifyFloat : public IRMutator {
     enum Strictness {
         FastMath,
@@ -58,6 +60,8 @@ public:
         any_strict_float |= (mode == Forced);
     }
 };
+
+}  // namespace
 
 bool strictify_float(std::map<std::string, Function> &env, const Target &t) {
     bool any_strict_float = false;
